Agentic_AI/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-openai
langgraph
pandas
openpyxl
fpdf
python-dotenv
streamlit